requirements.txt
bleach==1.5.0
captcha==0.2.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
flake8==3.5.0
Flask==0.12.2
h5py==2.7.1
html5lib==0.9999999
idna==2.6
imutils==0.4.3
ipython==6.2.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.9.6
Keras==2.0.8
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
nltk==3.2.5
numpy==1.13.3
olefile==0.44
pandas==0.20.3
parso==0.1.0
pew==1.1.0
pickleshare==0.7.4
Pillow==4.3.0
pipenv==8.3.1
prompt-toolkit==1.0.15
protobuf==3.4.0
psutil==5.3.1
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2014.7
PyYAML==3.12
gunicorn==19.7.1
requests==2.18.4
scikit-learn==0.19.1
scipy==0.19.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorflow==1.3.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
Werkzeug==0.12.2
win-unicode-console==0.5
